backend/requirements.txt
#!/bin/bash

# Create requirements.txt for backend
cat > /home/ubuntu/BattInsight/improved_structure/backend/requirements.txt << 'EOF'
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.7
psycopg2-binary==2.9.5
pandas==1.5.3
numpy==1.24.2
gunicorn==20.1.0
python-dotenv==1.0.0
openai==0.27.2
requests==2.28.2
openpyxl==3.1.2
xlsxwriter==3.0.9
pytest==7.3.1
pytest-flask==1.2.0
Flask-Limiter==3.3.0
Flask-Caching==2.0.2
